requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
jinja2==3.1.2
python-multipart==0.0.6
bcrypt==4.0.1
PyJWT==2.7.0